pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "test_automation_project"
version = "0.1.0"
description = "Test Automation Framework"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "pytest>=8.3.4",
    "selenium>=3.141.0", 
    "playwright>=1.52.0",
    "allure-pytest>=2.13.5",
    "appium-python-client>=1.3.0",
    "requests>=2.32.3",
    "mysql-connector-python>=9.2.0",
    "python-dotenv>=1.0.1",
    "webdriver-manager>=4.0.2",
    "pytest-html>=4.1.1",
]

[tool.pytest.ini_options]
testpaths = ["test_cases"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--tb=short"
markers = [
    "sanity: marks tests as sanity tests",
    "smoke: marks tests as smoke tests", 
    "regression: marks tests as regression tests",
    "flaky: marks tests as flaky (will retry on failure)",
] 